# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * STL Enhancer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StlStatsBoundingBox = {
  minX: number;
  minY: number;
  minZ: number;
  maxX: number;
  maxY: number;
  maxZ: number;
};

export interface StlStats {
  /** Number of triangles in the mesh */
  triangleCount: number;
  /** Number of vertices */
  vertexCount: number;
  boundingBox: StlStatsBoundingBox;
  /** Approximate volume */
  volume: number;
  /** Total surface area */
  surfaceArea: number;
  /** Whether the mesh is watertight/manifold */
  isManifold: boolean;
  /** Number of duplicate triangles found */
  duplicateTriangles: number;
  /** Number of degenerate (zero-area) triangles */
  degenerateTriangles: number;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export type EnhanceStlBody = {
  /** STL file to enhance */
  file: Blob;
  /**
   * Number of Laplacian smoothing passes
   * @minimum 0
   * @maximum 20
   */
  smoothingIterations?: number;
  /** Remove duplicate vertices and degenerate triangles */
  removeDuplicates?: boolean;
  /** Fix inverted/inconsistent normals */
  fixNormals?: boolean;
  /** Fill open boundary loops (holes) in the mesh */
  fillHoles?: boolean;
  /**
   * Maximum number of boundary edges in a hole to fill
   * @minimum 3
   * @maximum 5000
   */
  maxHoleSize?: number;
};

export type GetStlStatsBody = {
  file: Blob;
};